pyproject.toml
[project]
name = "pyconnect-lan-can"
version = "0.0.0"
description = "Zehnder ComfoConnect LAN C emulator speaking to an ESPHome ECB1 CAN bridge"
requires-python = ">=3.10"
dependencies = [
    "python-can>=4.3",
    "protobuf>=4.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyconnect_lan_can"]

[project.entry-points."can.interface"]
ecb1 = "pyconnect_lan_can.ecb1.bus:Ecb1Bus"
